Fix simulation error
The problem was that the time simulation (how long the animation should
take) was done by just waiting X milliseconds after each iteration.
Sadly this way only the last led layer would get the main activation time.
Now the time is properly sampled and the next state adjusted
accordingly.

Only sad thing about this is that now we can see that actual simulation
can't be done for more than 3 layers at a time, as the computational
load is too much for the arduino.